requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.0
openai>=2.23.0
google-genai>=1.66.0
aiofiles>=24.1.0
